full refactor sys0
