+++ contracts/proto0/mocks/RTokenMockP0.sol
@@ -3,8 +3,8 @@ pragma solidity 0.8.4;
 
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
-import "../test/ERC20Mock.sol";
-import "./interfaces/IStakingPool.sol";
+import "../../mocks/ERC20Mock.sol";
+import "../interfaces/IStakingPool.sol";
 
 interface IRToken is IERC20 {}
