+++ test/prod/RToken.test.ts
@@ -1,17 +1,17 @@
 import { ethers } from 'hardhat'
 import { expect } from 'chai'
-import { ZERO_ADDRESS, BN_SCALE_FACTOR } from '../common/constants'
-import { bn, divCeil } from '../common/numbers'
-import { advanceTime } from './utils/time'
+import { ZERO_ADDRESS, BN_SCALE_FACTOR } from '../../common/constants'
+import { bn, divCeil } from '../../common/numbers'
+import { advanceTime } from '../utils/time'
 import { BigNumber, BigNumberish, ContractFactory } from 'ethers'
 import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers'
-import { ERC20Mock } from '../typechain/ERC20Mock.d'
-import { CircuitBreaker } from '../typechain/CircuitBreaker.d'
-import { ReserveRightsTokenMock } from '../typechain/ReserveRightsTokenMock.d'
-import { RSR } from '../typechain/RSR.d'
-import { RTokenMock } from '../typechain/RTokenMock.d'
-import { TXFeeCalculatorMock } from '../typechain/TXFeeCalculatorMock.d'
-import { IBasketToken, IRTokenConfig, IRSRConfig, IRTokenParams } from '../common/configuration'
+import { ERC20Mock } from '../../typechain/ERC20Mock.d'
+import { CircuitBreaker } from '../../typechain/CircuitBreaker.d'
+import { ReserveRightsTokenMock } from '../../typechain/ReserveRightsTokenMock.d'
+import { RSR } from '../../typechain/RSR.d'
+import { RTokenMock } from '../../typechain/RTokenMock.d'
+import { TXFeeCalculatorMock } from '../../typechain/TXFeeCalculatorMock.d'
+import { IBasketToken, IRTokenConfig, IRSRConfig, IRTokenParams } from '../../common/configuration'
 
 // Sample Values for Configuration
 const stakingDepositDelay = 3600 // seconds
