+++ test/prod/RTokenDeployer.test.ts
@@ -1,16 +1,16 @@
 import { ethers } from 'hardhat'
 import { expect } from 'chai'
-import { expectInReceipt } from '../common/events'
-import { ZERO_ADDRESS } from '../common/constants'
+import { expectInReceipt } from '../../common/events'
+import { ZERO_ADDRESS } from '../../common/constants'
 import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers'
-import { RTokenMock } from '../typechain/RTokenMock.d'
-import { InsurancePoolMock } from '../typechain/InsurancePoolMock.d'
-import { RTokenDeployer } from '../typechain/RTokenDeployer.d'
-import { IBasketToken, IRSRConfig, IRTokenParams } from '../common/configuration'
-import { ReserveRightsTokenMock } from '../typechain/ReserveRightsTokenMock.d'
-import { RSR } from '../typechain/RSR.d'
+import { RTokenMock } from '../../typechain/RTokenMock.d'
+import { InsurancePoolMock } from '../../typechain/InsurancePoolMock.d'
+import { RTokenDeployer } from '../../typechain/RTokenDeployer.d'
+import { IBasketToken, IRSRConfig, IRTokenParams } from '../../common/configuration'
+import { ReserveRightsTokenMock } from '../../typechain/ReserveRightsTokenMock.d'
+import { RSR } from '../../typechain/RSR.d'
 import { ContractFactory } from 'ethers'
-import { CompoundMath } from '../typechain/CompoundMath.d'
+import { CompoundMath } from '../../typechain/CompoundMath.d'
 
 describe('RTokenDeployer contract', function () {
   let owner: SignerWithAddress
